How to Use
Youo need to have Gephi installed on your local device. https://gephi.org/users/download/

You also need a method to generate a gephi graph file from your csv dataset.
The python script runs on the data set locally in your drive.  It was explained in https://towardsdatascience.com/from-csv-to-github-pages-in-5-steps-publishing-an-interactive-social-network-of-the-marvel-7b8374bf44fb

If you wish to use a different csv-dataset, simply change the file name in the script so that it matches where your dataset is hosted.
The script will generate a file that is saved in a .gexf format file .

For now, you can simply use the one that is provided here,  trial-network(1).gexf .
If so, download it and then upload it into Gephi.
Once you are in Gephi, add in new plugins for the DAG analysis.

There are a number of ways to run Gephi for node analysis and edge analysis based on clustering.
The best feature on Gephi is to enlrage your mouse diameter  until you see the arrows emerge for the directed graph.

We can each play with this and generate our hearts happiness of DAG graphs with it!
Enjoy
~Hanna